+++ src/intrinsics/components/Acceleration.hpp
@@ -1,7 +1,9 @@
 #pragma once
 
 namespace Components {
+
+    //CGID: `:acceleration`
     struct Acceleration {
-        double d2x{0.0}, d2y{0.0}, d2z{0.0};
+        float d2x{0.0}, d2y{0.0}, d2z{0.0};
     };
 }
